requirements.txt
alembic==1.13.1
Authlib==1.6.6
beautifulsoup4==4.14.3
Flask==2.3.3
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
groq==1.0.0
gunicorn==21.2.0
psutil==7.2.1
pytest==9.0.2
python-dotenv==1.0.1
python-libnmap==0.7.3
requests==2.31.0
SQLAlchemy==2.0.45
Werkzeug==3.1.5